Custom slice type/date parsing

+++ handlers.go
@@ -7,27 +7,40 @@ import (
     "encoding/json"
     "strconv"
     "os"
+    "sort"
+    "time"
 
     "github.com/gorilla/mux"
 )
 
 const (
     DataFile string = "ping.data"
+    AllDevices string = "all"
+    DateLayoutISO string = "2006-01-02"
 )
 
 func StorePing(w http.ResponseWriter, r *http.Request) {
+    // get path params
     vars := mux.Vars(r)
     id := vars["deviceId"]
     time := vars["epochTime"]
-    t, _ := strconv.ParseUint(time, 0, 32)
+    t, _ := strconv.ParseInt(time, 0, 64)
 
-    var m map[string][]uint32;
+    // get data from file otherwise instantiate
+    var m map[string]int64arr;
     if PingMapCreated() {
    	m, _ = GetPingMap()
     } else {
-        m = make(map[string][]uint32)
+        m = make(map[string]int64arr)
     }
-    m[id] = append(m[id], uint32(t))
+
+    // append time to slice
+    m[id] = append(m[id], int64(t))
+
+    // sort slice
+    sort.Sort(m[id])
+
+    // write data to file
     b, _ := GetBytes(m)
     err := ioutil.WriteFile(DataFile, b, 0600)  
     if err != nil {
@@ -36,17 +49,33 @@ func StorePing(w http.ResponseWriter, r *http.Request) {
 }
 
 func RetrievePing(w http.ResponseWriter, r *http.Request) {
-    //vars := mux.Vars(r)
-    //id := vars["deviceId"]
-    //from := vars["from"]
+    // get path params
+    vars := mux.Vars(r)
+    id := vars["deviceId"]
+    from := vars["from"]
+
+    // parse from to ISO date
+    fromDate, err := time.Parse(DateLayoutISO, from) 
+    if err != nil {
+ 	fmt.Printf("Date parse error")
+    }
+    fmt.Printf("ISO date is %v", fromDate)
 
     m, err := GetPingMap()
     if err != nil {
 	fmt.Printf("Could not get ping map")
     }
-    //pings := m[id]
-    pingResponse, err := json.Marshal(m)
-    fmt.Fprintf(w, "%s", pingResponse)
+
+    if id == AllDevices {
+	// filter on every slice	
+    } else {
+        pings := m[id]
+        pingResponse, err := json.Marshal(pings)
+    	if err != nil {
+	    fmt.Printf("Failed to marshal to json")
+        }
+        fmt.Fprintf(w, "%s", pingResponse)
+    }
 }
 
 func RetrievePingTo(w http.ResponseWriter, r *http.Request) {
@@ -61,7 +90,6 @@ func RetrievePingTo(w http.ResponseWriter, r *http.Request) {
 }
 
 func RetrieveDevices(w http.ResponseWriter, r *http.Request) {
-    fmt.Fprintf(w, "Get all devices")
     devices, err := GetPingMapKeys()
     if err != nil {
 	fmt.Printf("Failed to get ping map keys.")
@@ -71,9 +99,11 @@ func RetrieveDevices(w http.ResponseWriter, r *http.Request) {
 }
 
 func DeleteAllPings(w http.ResponseWriter, r *http.Request) {
-    err := os.Remove(DataFile)     
-    if err != nil {
- 	fmt.Printf("Failed to delete ping data file.")
-    }   
+    if PingMapCreated() {
+        err := os.Remove(DataFile)     
+        if err != nil {
+ 	    fmt.Printf("Failed to delete ping data file.")
+        }   
+    }
 }
